Archive stale issues and pull request (#906)
Enable a bot that will automatically archive issues
older than 6 months or PR older than 2 months.

The bot will run every our and will archive one single
issue per hour.

Archived issues will be labeled as "archived" and can
be reopened if needed. Commenting on the issue or PR
or simply updating the PR branch is enough to count
as "activity" and reset the stale counter.